+++ ktorfit-ksp/src/main/kotlin/KtorfitProcessor.kt
@@ -4,9 +4,7 @@ import com.google.devtools.ksp.processing.*
 import com.google.devtools.ksp.symbol.KSAnnotated
 import com.google.devtools.ksp.symbol.KSFunctionDeclaration
 import de.jensklingenberg.ktorfit.generator.generateImplClass
-import de.jensklingenberg.ktorfit.generator.generateKtorfitExtClass
 import de.jensklingenberg.ktorfit.http.*
-
 import de.jensklingenberg.ktorfit.model.toClassData
 
 @AutoService(SymbolProcessorProvider::class)
@@ -41,10 +39,6 @@ public class KtorfitProcessor(private val env: SymbolProcessorEnvironment) : Sym
             }
 
         generateImplClass(classDataList, codeGenerator)
-
-        if (classDataList.isNotEmpty()) {
-           // generateKtorfitExtClass(classDataList, env.platforms.any { it.platformName == "JS" }, codeGenerator)
-        }
 
         return emptyList()
     }
