+++ ktorfit-lib/src/commonMain/kotlin/de/jensklingenberg/ktorfit/DefaultKtorfitService.kt
@@ -3,8 +3,5 @@ package de.jensklingenberg.ktorfit
 import de.jensklingenberg.ktorfit.internal.KtorfitClient
 
 class DefaultKtorfitService : KtorfitService {
-    override fun setClient(client: KtorfitClient) {
-
-    }
-
+    override fun setClient(client: KtorfitClient) {}
 }
